rgw: more headers cleanup
Dependency reduction
