+++ src/rgw/rgw_bucket.cc
@@ -13,6 +13,7 @@
 #include "common/errno.h"
 #include "common/ceph_json.h"
 #include "include/scope_guard.h"
+
 #include "rgw_rados.h"
 #include "rgw_zone.h"
 #include "rgw_acl.h"
