+++ src/rgw/rgw_log.h
@@ -3,10 +3,11 @@
 
 #ifndef CEPH_RGW_LOG_H
 #define CEPH_RGW_LOG_H
+
 #include <boost/container/flat_map.hpp>
 #include "rgw_common.h"
-#include "common/Formatter.h"
 #include "common/OutputDataSocket.h"
+#include "common/Mutex.h"
 
 class RGWRados;
 
@@ -110,13 +111,13 @@ struct rgw_log_entry {
     }
     DECODE_FINISH(p);
   }
-  void dump(Formatter *f) const;
+  void dump(ceph::Formatter *f) const;
   static void generate_test_instances(list<rgw_log_entry*>& o);
 };
 WRITE_CLASS_ENCODER(rgw_log_entry)
 
 class OpsLogSocket : public OutputDataSocket {
-  Formatter *formatter;
+  ceph::Formatter *formatter;
   Mutex lock;
 
   void formatter_to_bl(bufferlist& bl);
@@ -138,7 +139,7 @@ int rgw_log_op(RGWRados *store, RGWREST* const rest, struct req_state *s,
 void rgw_log_usage_init(CephContext *cct, RGWRados *store);
 void rgw_log_usage_finalize();
 void rgw_format_ops_log_entry(struct rgw_log_entry& entry,
-			      Formatter *formatter);
+			      ceph::Formatter *formatter);
 
 #endif /* CEPH_RGW_LOG_H */
