+++ src/rgw/rgw_object_expirer_core.h
@@ -28,17 +28,12 @@
 #include "include/utime.h"
 #include "include/str_list.h"
 
-#include "rgw_user.h"
-#include "rgw_bucket.h"
-#include "rgw_rados.h"
-#include "rgw_acl.h"
-#include "rgw_acl_s3.h"
-#include "rgw_log.h"
-#include "rgw_formats.h"
-#include "rgw_usage.h"
-
+class CephContext;
 class RGWSI_RADOS;
 class RGWSI_Zone;
+class RGWRados;
+class RGWBucketInfo;
+class cls_timeindex_entry;
 
 class RGWObjExpStore {
   CephContext *cct;
