+++ src/rgw/rgw_opa.cc
@@ -2,6 +2,7 @@
 // vim: ts=8 sw=2 smarttab
 
 #include "rgw_opa.h"
+#include "rgw_http_client.h"
 
 #define dout_context g_ceph_context
 #define dout_subsys ceph_subsys_rgw
