+++ src/rgw/rgw_rest_s3.cc
@@ -1313,7 +1313,7 @@ static inline void map_qs_metadata(struct req_state* s)
   for (const auto& elt : params) {
     std::string k = boost::algorithm::to_lower_copy(elt.first);
     if (k.find("x-amz-meta-") == /* offset */ 0) {
-      add_amz_meta_header(s->info.x_meta_map, k, elt.second);
+      rgw_add_amz_meta_header(s->info.x_meta_map, k, elt.second);
     }
   }
 }
