+++ src/tools/rbd/action/Perf.cc
@@ -22,6 +22,7 @@
 #include <boost/assign.hpp>
 #include <boost/bimap.hpp>
 #include <boost/program_options.hpp>
+#include "json_spirit/json_spirit.h"
 
 namespace rbd {
 namespace action {
